// src/models/User.ts
export interface User {
  id: string;
  name: string;
  firstName: string;
  lastName: string;
  email: string;
  profileImage?: string;
  organization?: string;
  facebook?: string;
  instagram?: string;
  linkedin?: string;
  youtube?: string;
  ageGroup?: string;
  website?: string;
  industry?: string;
  designation?: string;
  city?: string;
  country?: string;
  profileMessage?: string;
  aboutMe?: string;
}